chore(config): change port datatype from int to int16

+++ infrastructure/meilisearch/config.go
@@ -2,7 +2,7 @@ package meilisearch
 
 type Config struct {
 	Host              string `yaml:"host"`
-	Port              int    `yaml:"port"`
+	Port              uint16 `yaml:"port"`
 	Timeout           int16  `yaml:"timeout_in_ms"`
 	DefaultCollection string `yaml:"default_collection"`
 	APIKey            string
